[Lock] remove Factory for LockFactory
