+++ src/Symfony/Component/Lock/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+5.0.0
+-----
+
+* `Factory` has been removed, use `LockFactory` instead.
+
 4.4.0
 -----
