+++ src/Symfony/Component/Lock/LockFactory.php
@@ -11,12 +11,43 @@
 
 namespace Symfony\Component\Lock;
 
+use Psr\Log\LoggerAwareInterface;
+use Psr\Log\LoggerAwareTrait;
+use Psr\Log\NullLogger;
+
 /**
  * Factory provides method to create locks.
  *
  * @author Jérémy Derussé <jeremy@derusse.com>
  * @author Hamza Amrouche <hamza.simperfit@gmail.com>
  */
-class LockFactory extends Factory
+class LockFactory implements LoggerAwareInterface
 {
+    use LoggerAwareTrait;
+
+    private $store;
+
+    public function __construct(StoreInterface $store)
+    {
+        $this->store = $store;
+
+        $this->logger = new NullLogger();
+    }
+
+    /**
+     * Creates a lock for the given resource.
+     *
+     * @param string     $resource    The resource to lock
+     * @param float|null $ttl         Maximum expected lock duration in seconds
+     * @param bool       $autoRelease Whether to automatically release the lock or not when the lock instance is destroyed
+     *
+     * @return Lock
+     */
+    public function createLock(string $resource, ?float $ttl = 300.0, bool $autoRelease = true)
+    {
+        $lock = new Lock(new Key($resource), $this->store, $ttl, $autoRelease);
+        $lock->setLogger($this->logger);
+
+        return $lock;
+    }
 }
